share task request struct
