+++ crates/shared/src/models/mod.rs
@@ -4,4 +4,5 @@ pub mod heartbeat;
 pub mod invite;
 pub mod metric;
 pub mod node;
+pub mod storage;
 pub mod task;
